% Auto-generated: do not edit by hand
\name{creditCard}

\alias{creditCard}

\title{CreditCard component}

\description{
CreditCard component provides Beautiful credit cards for your payment forms See react component here: https://github.com/amarofashion/react-credit-cards
}

\usage{
creditCard(id=NULL, cvc=NULL, expiry=NULL, focus=NULL, name=NULL,
number=NULL, locale=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{cvc}{Character. CSV on credit card.}

\item{expiry}{Character. expiry date on credit card.}

\item{focus}{Character. focus on credit card entry}

\item{name}{Character. Name as it appears on credit card}

\item{number}{Character. account number credit card}

\item{locale}{Named list. localization text - like the words 'valid thru' on card}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

